+++ @xen-orchestra/xapi/cli.js
@@ -4,5 +4,5 @@
 
 const { Xapi } = require('./')
 require('xen-api/dist/cli.js')
-  .default(opts => new Xapi({ ignoreNobakVdis: true, ...opts }))
+  .default(opts => new Xapi(opts))
   .catch(console.error.bind(console, 'FATAL'))