+++ packages/xen-api/src/_parseUrl.js
@@ -1,4 +1,4 @@
-const URL_RE = /^(?:(https?:)\/*)?(?:([^:]+):([^@]+)@)?(?:\[([^\]]+)\]|([^:/]+))(?::([0-9]+))?(\/[^?#]*)?$/
+const URL_RE = /^(?:(https?:)\/*)?(?:(([^:]*)(?::([^@]*))?)@)?(?:\[([^\]]+)\]|([^:/]+))(?::([0-9]+))?(\/[^?#]*)?$/
 
 export default url => {
   const matches = URL_RE.exec(url)
@@ -6,12 +6,21 @@ export default url => {
     throw new Error('invalid URL: ' + url)
   }
 
-  const [, protocol = 'https:', username, password, ipv6, hostname = ipv6, port, pathname = '/'] = matches
-  const parsedUrl = { protocol, hostname, port, pathname }
-  if (username !== undefined) {
+  const [, protocol = 'https:', auth, username = '', password = '', ipv6, hostname = ipv6, port, pathname = '/'] =
+    matches
+  const parsedUrl = {
+    protocol,
+    hostname,
+    port,
+    pathname,
+
+    // compat with url.parse
+    auth,
+  }
+  if (username !== '') {
     parsedUrl.username = decodeURIComponent(username)
   }
-  if (password !== undefined) {
+  if (password !== '') {
     parsedUrl.password = decodeURIComponent(password)
   }
   return parsedUrl