+++ packages/xen-api/src/cli.js
@@ -44,13 +44,14 @@ const usage = 'Usage: xen-api <url> [<user> [<password>]]'
 
 async function main(createClient) {
   const opts = minimist(process.argv.slice(2), {
-    string: ['session-id', 'fallback-addresses'],
+    string: ['proxy', 'session-id'],
     boolean: ['allow-unauthorized', 'help', 'read-only', 'verbose'],
 
     alias: {
       'allow-unauthorized': 'au',
       debounce: 'd',
       help: 'h',
+      proxy: 'p',
       'read-only': 'ro',
       verbose: 'v',
     },
@@ -82,18 +83,14 @@ async function main(createClient) {
     })
   }
 
-  let fallBackAddresses = []
-  if (opts['fallback-addresses']) {
-    fallBackAddresses = opts['fallback-addresses'].split(',')
-  }
   const xapi = createClient({
     url: opts._[0],
     allowUnauthorized: opts.au,
     auth,
     debounce: opts.debounce != null ? +opts.debounce : null,
+    httpProxy: opts.proxy,
     readOnly: opts.ro,
     syncStackTraces: true,
-    fallBackAddresses,
   })
   await xapi.connect()