+++ packages/xo-web/src/common/xo/migrate-vm-modal/index.js
@@ -2,6 +2,7 @@ import BaseComponent from 'base-component'
 import every from 'lodash/every'
 import find from 'lodash/find'
 import forEach from 'lodash/forEach'
+import isEmpty from 'lodash/isEmpty'
 import map from 'lodash/map'
 import React from 'react'
 import store from 'store'
@@ -81,21 +82,16 @@ export default class MigrateVmModalBody extends BaseComponent {
     )
 
     this._getTargetNetworkPredicate = createSelector(
-      createPicker(
-        () => this.props.pifs,
-        () => this.state.host.$PIFs
-      ),
-      pifs => {
-        if (!pifs) {
-          return false
-        }
-
-        const networks = {}
-        forEach(pifs, pif => {
-          networks[pif.$network] = true
+      () => this.props.networks,
+      () => this.state.host.$poolId,
+      (networks, poolId) => {
+        const _networks = {}
+        forEach(networks, network => {
+          if (network.$poolId === poolId) {
+            _networks[network.id] = true
+          }
         })
-
-        return network => networks[network.id]
+        return isEmpty(_networks) ? false : network => _networks[network.id]
       }
     )